+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Remove maximum character limit for phone numbers
 - Add diabetes treatment question in medical history screen
 - Hide resend sms button when OTP attempts are blocked
+- Show patient died status in patient summary screen
 - [In Progress: 19 Aug 2021] UI improvements for medical history screen
   - Show hypertension diagnosis and treatment in single card
   - Show separate cards for hypertension and diabetes diagnosis